+++ chopper/lib/src/request.dart
@@ -108,10 +108,6 @@ class PartValue<T> {
     this.value,
   );
 
-  @Deprecated('Prefer copyWith method')
-  PartValue<NewType> replace<NewType>({String? name, NewType? value}) =>
-      copyWith<NewType>(name: name, value: value);
-
   /// Makes a copy of this PartValue, replacing original values with the given ones.
   /// This method can also alter the type of the request body.
   PartValue<NewType> copyWith<NewType>({String? name, NewType? value}) =>