+++ chopper/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Remove deprecated `FileField`, use `PartFile` instead
 - Remove deprecated `Request.replace`, use `Request.copyWith` instead
+- Remove deprecated `PartValue.replace`, use `PartValue.copyWith` instead
+
 
 ## 4.0.0-nullsafety.0